--- projects/plugins/jetpack/extensions/blocks/author-recommendation/editor.scss
@@ -1,9 +1,0 @@
-/**
- * Editor styles for Author-recommendation
- */
-
-.wp-block-jetpack-author-recommendation {
-    img.icon {
-        border-radius: 50%;
-    }
- }
