--- projects/plugins/jetpack/extensions/blocks/author-recommendation/test/fixtures/jetpack__author-recommendation.html
@@ -1,1 +1,0 @@
-<!-- wp:jetpack/author-recommendation /-->
