+++ projects/plugins/jetpack/extensions/blocks/blogroll/block.json
@@ -1,8 +1,8 @@
 {
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 3,
-	"name": "jetpack/author-recommendation",
-	"title": "Author Recommendation (Beta)",
+	"name": "jetpack/blogroll",
+	"title": "Blogroll (Beta)",
 	"description": "Select the sites you follow and share them with your users.",
 	"keywords": [],
 	"version": "12.5.0",
