+++ projects/plugins/jetpack/extensions/blocks/blogroll/editor.scss
@@ -1,0 +1,9 @@
+/**
+ * Editor styles for Blogroll
+ */
+
+.wp-block-jetpack-blogroll {
+    img.icon {
+        border-radius: 50%;
+    }
+ }
