+++ projects/plugins/jetpack/extensions/blocks/blogroll/icon.js
