+++ projects/plugins/jetpack/extensions/blocks/blogroll/style.scss
@@ -1,4 +1,4 @@
-.wp-block-jetpack-author-recommendation {
+.wp-block-jetpack-blogroll {
     display: flex;
     flex-direction: column;
     gap: calc(8px);
