+++ projects/plugins/jetpack/extensions/blocks/blogroll/test/fixtures/jetpack__blogroll.html
@@ -1,0 +1,1 @@
+<!-- wp:jetpack/blogroll /-->
