+++ projects/plugins/jetpack/extensions/blocks/blogroll/test/fixtures/jetpack__blogroll.parsed.json
@@ -1,6 +1,6 @@
 [
 	{
-		"blockName": "jetpack/author-recommendation",
+		"blockName": "jetpack/blogroll",
 		"attrs": {},
 		"innerBlocks": [],
 		"innerHTML": "",
