+++ projects/plugins/jetpack/extensions/index.json
@@ -56,7 +56,7 @@
 	"beta": [
 		"ai-chat",
 		"amazon",
-		"author-recommendation",
+		"blogroll",
 		"google-docs-embed",
 		"launchpad-save-modal",
 		"recipe",
